+++ src/main/java/vazkii/aurelienribon/tweenengine/primitives/MutableInteger.java
@@ -16,10 +16,10 @@ public class MutableInteger extends Number implements TweenAccessor<MutableInteg
 		this.value = value;
 	}
 
-	@Override public int intValue() {return (int) value;}
-	@Override public long longValue() {return (long) value;}
-	@Override public float floatValue() {return (float) value;}
-	@Override public double doubleValue() {return (double) value;}
+	@Override public int intValue() {return value;}
+	@Override public long longValue() {return value;}
+	@Override public float floatValue() {return value;}
+	@Override public double doubleValue() {return value;}
 
 	@Override
 	public int getValues(MutableInteger target, int tweenType, float[] returnValues) {
